Compile and link integrator/predictor sources instead of #include "*.cc"
Several .cc files were #included into the integrators rather than
compiled once and linked. This only worked because each affected target
had a single translation unit; adding a second TU that included any of
them would break the link with duplicate symbols.

Split them into declaration headers and linked sources, following the
existing extent.cc pattern:

- background, kabsch, sigma_estimation: include/integrator/*.hpp +
  src/integrator/*.cc (per-target sources, matching
  extent/lp_correction/ coordinate_system).
- predict.cc (predict, ray_predictors, utils, threadpool, with
  index_generators kept header-only as it is fully inline): a new
  `predictor` static library, linked by baseline_predictor,
  baseline_integrator and the GPU integrator.

Drops the now-unused ../baseline/integrator and ../baseline/predictor
include paths from the integrator targets and their tests.
