+++ baseline/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_subdirectory(spotfinder)
 add_subdirectory(refiner)
 add_subdirectory(indexer)
-add_subdirectory(integrator)
 add_subdirectory(predictor)
+add_subdirectory(integrator)
